[hack] whitelist $JIRI_ROOT from deletion when jiri fails
This patch prevent jiri from deleting $JIRI_ROOT if jiri update fails
due to the hack introduced in the last patch.

+++ project/operations.go
@@ -180,8 +180,10 @@ func (op createOperation) Run(jirix *jiri.X) (e error) {
 	}
 
 	if err := op.checkoutProject(jirix, cache); err != nil {
-		if err := os.RemoveAll(op.destination); err != nil {
-			jirix.Logger.Warningf("Not able to remove %q after create failed: %s", op.destination, err)
+		if op.destination != jirix.Root {
+			if err := os.RemoveAll(op.destination); err != nil {
+				jirix.Logger.Warningf("Not able to remove %q after create failed: %s", op.destination, err)
+			}
 		}
 		return err
 	}
